Extra explanations regarding stack arguments and stack growth.
Add the note that explain the calling convention and that the argc and argv are not actually taken from the stack.
Add explanations regarding the possible way to make valid stack accesses, when stack's memory is lazily allocated.
Add a Kile project to be able to manage easier all project's Latex files.